requirements.txt
boto3==1.35.38
botocore==1.35.38
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1